refactor: 새로 정의한 타입으로 import하도록 수정 / DP-318
