+++ src/features/map/components/MapOverlay/MapPlayerOverlay/MapPlayerOverlay.tsx
@@ -1,5 +1,5 @@
 import Button from '@/components/Button/Button';
-import type { PlayerOverlayData } from '@/features/map/types/player';
+import type { PlayerOverlayData } from '@/features/map/schemas/overlaySchema';
 
 import styles from '../MapOverlay.module.scss';
